Move ResultPrinter, refs 2740
